backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ai-edge-litert>=2.1.3",
    "fastapi>=0.135.1",
    "numpy>=2.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn>=0.42.0",
]
